MAINT: move `import_array` before module creation in module inits
This seems to be the right order, fixes an issue where the reference
to the newly created module isn't decreased if `import_array` fails.

Note that `import_array` already does its own error checking, and
returns NULL if NumPy initialization fails, so no checks are needed.
This makes the code a bit more concise.

Closes gh-16231