MAINT: add a more informative error message for broken installs
The line:
```
ImportError: cannot import name '_ccallback_c' from 'scipy._lib'
```
can be found in a lot of bug reports floating around on Stack Overflow
and in other places. It has nothing to do with `_ccallback_c` in
all cases I looked at, it's always a generic problem with extension
modules. The solution is typically to reinstall, and the problem
goes away. So say that in the `ImportError` message.

Also seen in gh-16300 just now.