CI: test distro Python install on Ubuntu Jammy (22.04 LTS)
This fixes the intent of this CI job, which was first added in
gh-16309. Builds on Debian/Ubuntu are "special", because those
distros patch `distutils`/`sysconfig` in ways that are quite
unhelpful. The build was broken until `meson-python` 0.6.0

We want to keep this job running to guard against regressions.
It's pretty quick to run, so costs little.

Closes gh-16300

[skip azurepipelines] [skip azp]