DOC: update guidance on a Meson FAQ entry
The answer speaks about working with the Meson maintainers to make the
warnings less scary. This has happened, so the guidance is outdated.

Keep the FAQ entry, because it is still relevant as long as the oldest
supported Meson build requirement is less than the version in which the
warning disappeared. Instead, amend the guidance to state that upgrading
Meson will, in fact, make it disappear.