MAINT: remove last (already safe) usage of `mktemp`
`mktemp` is deprecated in the `tempfile` docs, because it's in
many cases insecure. This last usage wasn't insecure, as noticed in the
code comment, however static checkers for security issues don't
understand that comment and will still trigger on `mktemp`.
So get rid of it to avoid future (false) security-related
communication.

Note that this is a follow-up to gh-3289