MAINT: future-proof `stats.kde` for changes in numpy casting rules
The issue is caused by inconsistent float128 handling. In this line:
```
result = gaussian_kernel_estimate[spec](self.dataset.T, self.weights[:, None],
                                        points.T, self.inv_cov, output_dtype)
```
the first 3 input arrays are float128, but `inv_cov` is float64
because `scipy.linalg` functions happen to return `float64` output for
`float128` input. `numpy.linalg` on the other hand will raise an
exception (which is the more sensible behavior).

Now why this changes with the NEP 50 branch and setting
`NPY_PROMOTION_STATE=weak` rather than the default (= legacy) is because
an expression of the form:
```
np.array([[3.5]], dtype=np.float64) / np.float128(3)
```
starts returning a float128 output rather than a float64 one. This
happens because of rule number 2 in NEP 50:

    NumPy scalars and 0-D arrays should behave consistently with their
    N-D counterparts.

So that is working as designed, it is just exposing some rather poor
dtype handling in `stats.kde` that happened to work before.