ENH: Use quick_select for PyArray_OrderFilterND instead of qsort.
Closes #1345.

Tests pass, but I haven't run benchmarks, so I don't know if this is
practiacally faster as well as theoretically faster.  I can't think
of a reason for a slowdown in PyArray_OrderFilterND, but the median
filtering code redirects comparisons through a function pointer
instead of using '<' with float, double, and unsigned byte arguments.
The compiler might decide to inline everything and keep the speed,
or keep the size down and accept the speed hit.