ENH: stats: Implement _munp() for the skewnorm distribution.
The skew-normal distribution has a moment-generating function from
which we can derive the noncentral moments.  This allows _munp(order, a),
the method of the rv_continuous class that computes the "standard" (i.e.
loc=0, scale=1) noncentral moment, to be implemented.  This is much
faster and more accurate than the default method, which uses
scipy.integrate.quad to compute the moments.