add an hint on what to use in case of matlab v7.3
Many new python users that move to python from matlab will need to load files from matlab to python, first thing that they see in google is to use sio.loadmat to load matlab files, which works sometimes. But when it does not work, I think a hint to use h5py will increase the probability for them to continue their move to python.