BLD: raise an error with clear message for too-new Python version
This works for both `python install xxx`, which gives:
```
RuntimeError: Your Python version is too new. SciPy 1.9 supports Python 3.8-3.11; if you are trying to build from source for the most recent SciPy version you may hit this error as well. Please build from the `main` branch on GitHub instead.
```
as the last line.

For installing through `meson` or `meson-python`, we get:
```
meson.build:55:2: ERROR: Problem encountered: Your Python version is too new. SciPy 1.9 supports Python 3.8-3.11; if you are trying to build from source for the most recent SciPy version you may hit this error as well. Please build from the `main` branch on GitHub instead.

A full log can be found at /home/path/to/log
Meson build setup failed! ({0} elapsed)
```

Through pip or build we get some more boilerplate at the end that is
not avoidable, but the error message is still clear:
```
      Program python3 found: YES (/home/rgommers/anaconda3/envs/scipy-dev/bin/python3.10)

      ../../meson.build:55:2: ERROR: Problem encountered: Your Python version is too new. SciPy 1.9 supports Python 3.8-3.11; if you are trying to build from source for the most recent SciPy version you may hit this error as well. Please build from the `main` branch on GitHub instead.

      ...
      AttributeError: module 'mesonpy' has no attribute 'prepare_metadata_for_build_wheel'

      During handling of the above exception, another exception occurred:

          ....
          raise CalledProcessError(retcode, cmd)
      subprocess.CalledProcessError: Command '['meson', 'setup', '--native-file=/home/rgommers/code/scipy/.mesonpy-native-file.ini', '-Ddebug=true', '-Doptimization=2', '--prefix=/home/rgommers/anaconda3/envs/scipy-dev', '/home/rgommers/code/scipy', '/home/rgommers/code/scipy/.mesonpy-_3w2qa5b/build']' returned non-zero exit status 1.

  note: This error originates from a subprocess, and is likely not a problem with pip.
error: metadata-generation-failed

× Encountered error while generating package metadata.
╰─> See above for output.

note: This is an issue with the package mentioned above, not pip.
hint: See above for details.
```

Closes gh-14986

Note that this patch should be reverted after SciPy 1.9.0 is released.
We just need a single sdist on PyPI with this message. After that
we can simply rely on the correct metadata in `pyproject.toml`.