BUG: optimize: Fix differential_evolution error message.
When DifferentialEvolutionSolver.solve() gets to the end of its
calculation and discovers that the constraints are not satisfied, it
sets the `success` attribute of the result to False and assigns a
string to the `message` attribute.  The code that generated the
message had a bug; in fact, it *should* have raised an exception if
it ever ran, but apparently the differential evolution code catches
and ignores such exceptions (a strange behavior that hides bugs like
this one).