BLD: update `meson` and `meson-python` versions for 1.9.0 release
This works around an issue where `pip install .` or `python -m build`
end up with a `py3-none` wheel tag, rather than the correct ABI tag.

Pinning the Meson version in `pyproject.toml` will only affect wheel
building, and should be fine.

For the change to `meson.build`, the upper bound doesn't seem to work.
This was never tested before, and is safe to remove.