CI: add Meson version number in environment.yml to rebuild Docker image (#16706)
This Docker image with all development dependencies only gets rebuilt
on a change to `environment.yml`. So if we bump the minimum version
in `meson.build` only, there's a problem. Hence adding the version
number here, so grepping the codebase finds this during a
version bump process.

This fixes the Gitpod image build in CI, it failed on:

    ERROR: Meson version is 0.62.2 but project requires >= 0.63.0

[skip azp]