BLD: install missing `.pxd` files, and update TODOs/FIXMEs in meson.build files (#16749)
* BLD: remove a warning suppression flag for HiGHS extension

This is no longer necessary, the PR linked in the removed comment
is included in our `highs` git submodule now.

* BLD: remove an outdated FIXME from the scipy/integrate/meson.build

This was fixed I believe (there are no skipped or xfailed tests either),
and the comment isn't very informative.

* BLD: remove another FIXME in `cython_optimize/meson.build`, it's not broken

The code can be simplified once the upstream feature is added, but
that's not a FIXME.

* BLD: remove comments about c++14/c++11 support

We can rely on C++14 unconditionally since 2021, see
http://scipy.github.io/devdocs/dev/toolchain.html#c-language-standards

* BLD: update info on including all license files as separate files

* BLD: install missing `scipy/*.pxd` files

No tests were failing, so something is fairly badly wrong in
the test suite.

Also add `optimize.cython_optimize` to the list of submodules that are
public API.

* BLD: remove FIXME about version string generation

This was fixed when adding `--git-dir` to the git invocation,
that made the result independent from the current directory
when invoking the build.

* BLD: remove TODO about `-fallow-argument-mismatch`

This flag is added implicitly via 'fortran_std=legacy' in the
top-level `meson.build`. Warnings that show up are silenced
in multiple modules, so there's no need for a separate TODO here.

* BLD: remove TODO for UNU.RAN version string

This isn't a very useful today, given that we don't use this
version string. The corresponding `setup.py` file also does the
same and doesn't have a TODO.

* BLD: update FIXME to a TODO in `linalg/meson.build`