BUG: stats.rv_histogram: address issue with non-uniform bins (#15381)
The documentation of `rv_histogram` was ambiguous about what it expected as input when bin sizes vary: counts or density? Adding to the confusion, the default behavior of `rv_histogram` was to assume that the input was a density, whereas the default of `np.histogram` is to produce bin counts. To resolve this, this PR adds a parameter `density` so that the user can specify whether the input is to be treated as counts or probability density. Because the default is different from that of `np.histogram`, this also warns the user to pass `density` explicitly.

Co-authored-by: Matt Haberland <mhaberla@calpoly.edu>