ENH: stats: Add the function odds_ratio. (#13340)
* ENH: stats: Add the function odds_ratio.

The new function `scipy.stats.contingency.odds_ratio` computes
the odds ratio for a 2x2 contingency table.  An option
allows the user to select either the sample odds ratio or the
conditional maximum likelihood estimate of the odds ratio.  The
returned object provides a method for computing the confidence
interval of the odds ratio.

Co-authored-by: Matt Haberland <mhaberla@calpoly.edu>