BUG/TST: linalg: Check the results of 'schur' more carefully.
The Schur decomposition is not unique, so testing the result
of linalg.schur against hard-coded matrices is not reliable.
The updated tests check that the desired properties of the
result are satisfied, and that the order of the diagonal
elements of the triangular matrix agrees with the given
'sort' parameter.

Closes gh-14517