MAINT: Unnecessary all-true array multiplication.
While testing openrefactory, I came across a SciPy report:

In particular:

 - https://pypi.openrefactory.com/scipy/62f53252fe6732ccd95a45ca#Inappropriate%20Logic-27M

 > In file: _distn_infrastructure.py, method: isf, a logical equality
 > check operation was performed with the same operand on both sides. The
 > comparison operation always returns either true or false. Such logical
 > short circuits in code lead to unintended behavior. iCR suggested that
 > the logical operation should be reviewed for correctness.

In particular the report is not totally accurate as `cond == cond` is a
quick way to create an all-True boolean array of the same size as cond,
and potentially quickly broadcast another array of different size.

Though in this particular 2 cases it seem that `q` and `cond`, or
`condX` and `cond` are anyway the same size, so my understanding is that
multiplying by `cond == cond` in those case is not necessary.