MAINT: Label new PRs with some modified files' languages
Maintainers and some contributors might be interested in
browsing PRs based on the languages they modify.
The "Cython", "Fortran" and "C/C++" exist to do so, but
have to be manually added during triage work.

This commit add a GitHub workflow to automatically label
Pull Requests with the "Cython", "Fortran" and "C/C++"
labels if they modify files whose names match one of the
associated defined globs.

The setup is inspired from scikit-learn's, which relies on
@thomasjpfan's work.

Co-authored-by: Thomas J. Fan <thomasjpfan@gmail.com>