DOC, TST: clarify Voronoi Qz (#16788)
Fixes #16773

* About ten years ago in 5bdd16ddbc, Pauli
matched upstream behavior by explicitly accounting
for the internal point at infinity that `Qhull` adds
when the `Qz` option is used with `Voronoi` (which,
for us, it is by default). That means that the Voronoi
`regions` data structure will typically have one extra
empty list for a Voronoi region of a point at infinity.
Correspondingly, the `point_region` data structure will have one
less element than `regions` in these `Qz` cases because there
you are explicitly finding the correspondence to the input
points/generators only.

* This is admittedly a bit confusing, so I've tried to clarify
the docs on this matter. To be honest I was tempted to change
the behavior to be less surprising ("principle of least surprise"
or something?), but thought that this wasn't quite a bug, and was
intentionally set ten years ago to match "upstream." I'm open to
changing it on my end, but the doc-only "fix" is my first suggestion
anyway.

* unrelated doc fix: `point_region` is an array, not a list, so
fix that (it is already fixed in another docstring, but not the
main one that gets exposed)

[ci skip]