BLD: suppress build warnings for `id_dist`
This is old F77 code that is quite hard to fix, and probably not
worth spending the effort on to fix (the main issue is in the
vendored FFTPACK). There are modernized copies of FFTPACK floating
around, but they all still have the issue that is the problem
here (REAL(8) to INTEGER(4) conversions).