BUG: linalg: Simplify test_sgesdd_lwork_bug_workaround() (#16965)
The test of the _compute_lwork() work-around for a bug in
LAPACK versions older than 3.10.1 is simplified.  Instead
of calling svd() in a subprocess with a large input array
and checking that it doesn't crash Python, we check that
the function sgesdd_lwork() returns the correct value.  It
is the incorrect value that resulted in the Python crash,
so this is a simpler test of the same issue.