BLD: make MKL detection more robust, add notes on TODOs
If the detection happens via CMake, that results in an upper-case
result for `blas.name()`. So deal with any case, and also that
of using `-Dblas=mkl_rt`.

Note that this does not fix things for using CMake, because that
defaults to the ILP64 interface (xref gh-16988).
Add notes on what is left for MKL support

Also add g77 ABI wrapper and fix a typo in f2py signature file in `_propack`
Note that this is an incomplete fix, because PROPACK is in a bad state,
as discussed in the issue linked in the code comment.