ENH: support complex functions in integrate.quad
This introduces support for complex valued integrands
in the integreate.quad method.  The integrand is
treated as real or complex based on the value of
a boolean flag introduced to the method's call
signature.

This enhancement is motivated by the need to support
complex functions in the expectation of circular
distributions in scipy.stats, See #16907.