TST: skip `Test_SVDS.test_small_sigma2` for float32 with PROPACK
This has been failing in CI for a while. There are lots of PROPACK
issues, so we'll just add one more skip for now. This test is
pretty sensitive, regular float32 tests all pass.

See gh-17004 for more details.

Also note that the skip condition had a problem, so ARPACK tests
would be skipped if PROPACK was not available - this is fixed now.