BUG: fix `signal.sosfilt` issue with complex dtypes and Intel compiler
This solves 12 test failures (see gh-17075) where the result for
complex64 and complex128 were wildly wrong when SciPy is built with the
`icc` compiler.

The code is very obscure, so it's not immediately clear whether it's
a compiler bug or code that was trying to be too smart. I suspect it's
the latter. For reference here is the Cython-generated C code:

Old code:
```C
      /* "scipy/signal/_sosfilt.pyx":46
 *         for n in xrange(n_samples):
 *
 *             x_cur = x[i, n]             # <<<<<<<<<<<<<<
 *
 *             for s in xrange(n_sections):
 */
      __pyx_t_8 = __pyx_v_i;
      __pyx_t_9 = __pyx_v_n;
      __pyx_v_x_cur = (*((float *) ( /* dim=1 */ ((char *) (((float *) ( /* dim=0 */ (__pyx_v_x.data + __pyx_t_8 * __pyx_v_x.strides[0]) )) + __pyx_t_9)) )));
```

New code:
```C
 *             x_cur = 1.0 * x[i, n]             # <<<<<<<<<<<<<<
 */
      __pyx_t_8 = __pyx_v_i;
      __pyx_t_9 = __pyx_v_n;
      __pyx_v_x_cur = (1.0 * (*((float *) ( /* dim=1 */ ((char *) (((float *) ( /* dim=0 */ (__pyx_v_x.data + __pyx_t_8 * __pyx_v_x.strides[0]) )) + __pyx_t_9)) ))));
```

[skip circle]

fix dtype issue