MAINT: pyproject.toml: Update build system requirements
Require "meson-python>=0.9.0", "Cython>=0.29.32", "pybind11>=2.10.0" and "pythran>=0.12.0" to build SciPy. These updated build-system requirements allow for consistent builds. The updated meson-python and Cython helps with Python 3.11 builds.

Probably a backport candidate to the 1.9 branch if Python 3.11 wheels are to be released on this branch.

(note that build-system requirements are only needed to build SciPy from source, not to run it after a regular pip install, and thus can be set fairly aggressive).

[wheel build]