Fixed bug in test_milp_timeout_16545 (#17137)
Previously, the test would check to see if res.message
started with the variable msg. However, it seems that
res.message frequently didn't start with msg, causing
an assertion error. To resolve this, I switched it
to only check and see if res.message is not None.

While this doesn't tell us if res.message is the correct
message, it also doesn't have to, since test_result does
that check as well. Since that function doesn't throw any
errors, it's safe to say that res.message is correct in
test_milp_timeout_16545 as well, even if we don't check.

Oddly, I wasn't able to change the inputs of the function,
so it still needs to take msg as input, even though msg isn't
actually used now. This isn't ideal, but since this code
is just being used for testing, it doesn't need to be elegant.