BLD: fix issue with incomplete threads dependency handling
All could using `#include <thread>` needs to have this `thread_dep`
dependency. HiGHS code was missing it, and for `stats/_qmc_cy.pyx`
it was incorrectly specified.

It's wrong in the distutils build as well, but let's ignore that
since it's much harder to change and we'll get rid of that build
soon.

See gh-17193 for the type of compile error this causes.