ENH: optimize.linprog: pass mip_rel_gap to HiGHS (#16897)
ENH: optimize.linprog: pass mip_rel_gap to HiGHS 

Also, when integrality is not None, update the `sol` dict with `mip_node_count`,
`mip_dual_bound`, and `mip_gap`.  Previously, it was updating the object
`res` which never got returned from the function

Finally, changed docstring in _linprog_highs, which previously incorrectly stated
that the mip_gap parameter returned is an absolute gap. It is a relative
gap according to highs documentation.