BUG: signal: Change types in the upfirdn utility function _output_len()
Change the types of the parameters of the Cython function `_output_len()`
from `np.intp_t` to the `np.int64_t`.  This avoids the potential
overflow that can occur if `np.intp_t` is 32 bit (e.g. 32 bit Linux
or Windows) and the input is very long, and also avoids a bug in
Cython 0.29.32 that occurs in 64 bit Windows.

Closes gh-17375.