BUG: integrate: simpson didn't handle integer n-d arrays.
A call such as `simpson([[1, 2, 3, 4, 5]], axis=-1)` would raise
an exception, because a calculation left an intermediate result
as an integer array, and then the code attempted to multiply
that result in-place by a floating point value.  Simply changing
a coefficient from 4 to 4.0 ensures that the array has a floating
point type.

Also added the missing "Returns" section to the `simpson` docstring.