MAINT: optimize.fmin_: fix wrap callbacks for these, too
Changing the way the callback is called in the minimize
methods changes the `fmin_` function, too. If the callback is
not wrapped, the user's callback of the `fmin_` functions will
get passed an OptimizeResult. Make sure that doesn't happen by
applying the same wrapper to the callbacks in the `fmin_`
functions.