BUG: add infeasibility checks to min_weight_full_bipartite_matching
The algorithm is already equipped with a handful of checks for
problem infeasibility, but the tests added here show two different
examples that were not caught by any of those checks. It is likely
possible to dive in and determine infeasibility in `_lapjvsp` itself, as
in both examples, the symptom is the algorithm hitting an infinite loop
(but a different infinite loop for each case), but here we simply add
the catch-all fix of checking that the problem is feasible before even
trying to solve it. This does come with a performance hit, but at least
the implementation of Hopcroft--Karp that we're using is fast.