BUG: correctly handle array-like types in scipy.io.savemat
Previously, the presence of a __dict__ attribute would unconditionally cause
savemat to treat the object as a mapping. This is problematic in the case of
custom array-like types, because by default *any* python class will have a
__dict__ attribute. This change makes it so that if an __array__ attribute is
present, the object will be treated as an array rather than as a mapping.