MAINT: Simplify masking of bad values in stats._multivariate.
Instead of relying on a custom _checkresult and manual broadcasting, we
can just use np.where and np.select as appropriate, which is simpler to
read (or at least not worse even if you don't know the semantics of
these functions); these also handle the broadcasting for us.

This will make all affected functions consistently return size-() arrays
for scalar inputs, which is a slight API change but actually more
consistent that the previous situation (multinomial would return a
size-() array for valid scalar inputs and a plain scalar for invalid
inputs; hypergeom would always return a plain scalar for scalar inputs).