Clarify input dimensions for directed_hausdorff method
The docstring stated that the method accepted 2-D arrays, which (combined with the example) led me to the misunderstanding that the input co-ordinates had to be in two spatial dimensions. Now I see that the input arrays are 2-D even for N-D coordinates. Hopefully the suggested change will help other people avoid having the same misunderstanding.

See #17756