ENH: stats: Add handling of censored data to univariate cont. distributions (#13699)
ENH: stats: Add handling of censored data to univariate cont. distributions.

This change allows the maximum likelihood estimation of univariate
continuous distribution parameters for censored data.

The censored data is represented with the new CensoredData class.
The class allows the represention of left-, right- and interval-censored
data.

The `fit` method is modified to check for the input `x` being an instance
of CensoredData and handled appropriately.