DOC: dev_quickstart: correct venv creation path
In the `pip+venv` directions of the "Basic workflow" section of the
"Contributor quickstart guide", the directions state "...execute the
following commands at the terminal from the base directory of your
SciPy clone", then direct users to run `python -m venv scipy-dev`.

This command creates a virtual environment within the clone, which
will cause build errors when running `meson` via `python dev.py
build`. The subsequent command, `source
$HOME/.venvs/scipy-dev/bin/activate` will not activate the virtual
environment just created; presumably that command is an artifact of
using `virtualenvwrapper`, as suggested in the "Using `virtualenv`"
section of the "Development environment guide (Ubuntu)".

This commit alters the code block in the `pip+venv` instructions
referenced above so that it creates a virtual environment in
`$HOME/.venvs/scipy-dev`, consistent with the `source` command on the
subsequent line. This alteration also ensures users avoid `meson`
errors when attempting to build the project in subsequent steps.

Signed-off-by: Geoffrey M. Oxberry <goxberry@gmail.com>
Signed-off-by: Geoffrey M Oxberry <goxberry@gmail.com>