STY: fix only staged files. (#17948)
Before, the staged tree was used only for finding the list of
files.  Now, linting is run on the staged version of the code as well.

If errors are detected, we recommend to the user that they run:

  ./tools/pre-commit.py --fix

That will execute the linter with the `--fix` flag, but only on the
files that were staged already.