BLD: use a relative path to numpy/pybind11/pythran include and lib dirs
This should make things work with in-tree virtual envs, as well as some
situations with Docker that had issues before.

The approach to determine absolute/relative paths unfortunately has to
be a bit convoluted, but I believe this should work in all
circumstances. A future improvement is still to make this work:
```
np_dep = dependency('numpy')
```
That way it won't be needed to run any Python code for the host
interpreter at build time (this is bad for cross-compiling).

Closes gh-16312