BLD: use meson-native dependency lookup for pybind11
This makes use of https://github.com/mesonbuild/meson/pull/11463 to find
pybind11 using the pybind11-config script installed in build isolation.

If that doesn't work, for example the version of Meson is too old, then
we fall back to the previous method of manually running some
introspection code, and no harm is done.