BLD: emit a warning when building from source on 32-bit Windows
We are getting semi-regular bug reports from users who are getting
builds from source and don't understand what is failing. They are
usually better served by installing 64-bit Python, so recommend that.

This will also make it easier for us to see whether a Windows user is
actually on 32-bit Python, because it can't be seen from the CPU info;
that is typically still x86-64. Unfortunately Python installers are
not labelled well enough. The 64-bit installer has a "recommended"
label, but it's fairly natural for users to grab, e.g.,
`python-3.10.9.exe` instead of `python-3.10.9-amd64.exe` from the FTP
download site.

xref gh-18143 for an example of a bug report due to this.