TST: fix issue with inaccurate `cython_blas` tests
These tests have been there since the initial introduction of
`cython_blas`. The `rtol` argument was set so high that anything would
pass (e.g., the `sasum` example compared 8.5 to 0 and it passed).

This also fixes the bug in the test routines themselves for
`sasum` and `snrm2`. Only test code was affected, not the corresponding
public functions.