BLD: require meson>=1.1.0, pybind11>=2.10.4, simplify pybind11 detection
Meson 1.1.0 supports `dependency('pybind11')` and that's now guaranteed
to work for pybind11 >= 2.10.4 (which adds the needed flag to
`pybind11-config` so Meson can use it). This change removes a bunch of
complexity, so it's worth defaulting to these fairly recent versions
of Meson and pybind11.

Also makes sure that pybind11 info is present in `scipy.show_config()`
output.