DEV: work around pathlib bug affecting dev.py for Python 3.9 on Windows
Also remove a workaround in a Windows job for this issue, and clean up
the code for `dev.py --win-cp-openblas` some more.

Note that this doesn't always work - in particular,
`openblas_support.py` will silently fail to do the right thing on
Windows if the git repo checkout is on the D: drive while
OpenBLAS gets downloaded to the C: drive. That is happening in a
GHA CI job, and it's plagued us on Azure before too (see gh-11967).