BLD: DOC: fix Sphinx doc build caching behavior for `.dev` versions
The regex for parsing `scipy.__version__` was off, causing Sphinx
to do a full rebuild each time you made a commit when working on
docs.

This change makes a commit a no-op for Sphinx caching, which reduces
doc build time from ~14 minutes to ~14 seconds.

With or without the changes in `conf.py`, changes to docstrings in
`.pyx` files or other non-Python files do not get picked up by
Sphinx' caching system (this is a sphinx/autodoc bug). When making
the caching behavior such that full rebuilds are less frequent,
then of course the chance of such docstrings not being reflected
in the html build output increases. Hence also add a `--no-cache`
flag to `dev.py doc` to force a full rebuild.

[skip actions] [skip cirrus]