BLD: fix yet another PROPACK build issue
The `setup.py` code is written kinda strangely, but checking the
build logs it is clear that the only OpenMP related flag is used
is `-U_OPENMP`, which undefines `_OPENMP` and hence the OpenMP
code in PROPACK is not used.

The define in the meson.build file was incorrectly done for C rather
than Fortran.

[skip cirrus] [skip circle]