DOC: avoid taking the log of something close to 0 in signal examples
Because the bohman & cosine windows are symmetrical, using `fft(window, 2048)`
would have 0 (or something very close to it) at the Nyquist frequency N/2=1024.
This is evident from a bit of manipulation of the definition of the discrete
Fourier transform. To avoid taking the log of something that's 0 (up to eps),
use an odd length for the fft, which ensures that the minimal absolute value
of `A / A.max()` in both examples rises from <1e-16 to ~5e-9.