ENH: lobpcg: remove unneeded normalization, default to in-place matrix multiplication (#17675)
The main algorithmic change in this PR is removing pre-scaling prior to
orthogonalization in the function `_b_orthonormalize`, which is key in
LOBPCG but may lead to numerical instabilities. Its input vectors are
often badly scaled, so the function needs scale-invariant Cholesky; see
https://netlib.org/lapack/lawnspdf/lawn14.pdf.

However, the need for normalization depends on how Cholesky is specifically
implemented. A separate unit test now checks that on all currently tested
platforms, Cholesky implementations are scale-invariant, i.e., normalization
is not needed.

Another significant change is to attempt performing as many in-place matmuls
as possible. Due to possible type mismatches, array slicing etc., in some
situation in-place matmuls are impossible and are still performed the old way.

Test coverage is improved across more dtypes, some complex cases are handled
better, and some test parameters are reduced to speed up things.