BUG: Fixes how the length of argument x of linalg.blas.?gbmv is checked
Given an m×n banded matrix A, linalg.blas.?gbmv does
y <- α A x + β y     if trans==0
y <- α A^T x + β y   if trans==1 (and A^H if trans==2)
Therefore x must be of length n when trans==0 and m otherwise.
This commit fixes the current code, which does the opposite.
Closes #18647.