ENH: CloughTocher2DInterpolator multiple times with different values (#18376)
Provide several hooks which a subclass can override. The use case is repeated evaluation of an interpolant with a fixed triangulation and `xi`, and varying `values`.

Note that all the additions here are --- strictly speaking --- private (their names start with a leading underscore), so SciPy makes no formal promises about API stability even if it's unlikely to change in a foreseeable future. Subclasses to make use of these hooks need to be developed and maintained downstream.