DOC: made sparse array/matrix docs only have array/matrix
Before, the docs for sparse arrays/matrices were written
for the array classes and "sparse array" was replaced with
"sparse matrix" to get the matrix class doc, but that did
not cover all strings that need to be replaced. Now, the
base docstrings have strings e.g. "{array|matrix}" that are
replaced with "array" or "matrix" to get the array or
matrix docs. Docstrings of the different sparse formats
are also made more consistent. See #18669.