BUG: trust-constr test throws expected warning
Previously, the unit test addressing the situation where trust-constr
returns status=2 for an infeasible solution failed due to warnings
related to a singular Jacobian and delta_grad == 0.0. Here, the unit
test is updated to ignore those warnings. Additionally, this commit
updates the documentation to describe how gtol is used as a tolerance
for constr_violation.