DOC: change sparse class method docstrings to "array/matrix"
Prior to this, the method docstrings of sparse classes
(and their parent classes) were a mix of "array"s and
"matrix"s. This commit makes this consistent by changing
all relevant instances of "array" and "matrix" to
"array/matrix". When sparse matrices are deprecated,
change all instances of "array/matrix" to "array".
See #18669.