DOC: from CONTRIBUTING.rst link to SciPy website
Links in CONTRIBUTING.rst went to other .rst files rather than
the SciPy website. This is problematic because the links in the
other .rst files don’t work unless they are rendered.

Furthermore, new contributors are likely to see CONTRIBUTING.rst
because it is linked from the new contributors popup.

This commit fixes the issue by linking to the SciPy website from
CONTRIBUTING.rst rather than other .rst files.

Also, the link to John Hunter's BSD pitch has suffered from link
rot, so I replaced the link to a new link to an archive of the
website.

Closes #18903.