DOC: indicate functions with multiple valid solutions in csgraph
Issue #17734 arose because an implementation change in Dijkstra's
algorithm led to a change in returned solution, in a case in which
multiple valid solutions were possible; the consequence was breaking
test cases in user code. This problem is common in algorithms in
csgraph. The purpose of this PR is to improve documentation to reduce
confusion by adding a note for each function that could have multiple
valid solutions.