DOC: move sparse class docs to array/matrix classes
Previously, the class docs of the sparse array/matrix
classes were defined in the base class, and then a string
replacement was used to obtain the array and matrix class
docs. To allow for more divergent array and matrix class
docs, it has been decided to simply define the docstrings
in the array/matrix classes, and this commit does that.
See #18669.